+++ src-tauri/src/lib.rs
@@ -12,6 +12,7 @@ pub fn run() {
     .plugin(tauri_plugin_dialog::init())
     .plugin(tauri_plugin_store::Builder::default().build())
     .plugin(tauri_plugin_os::init())
+    .plugin(tauri_plugin_clipboard_manager::init())
     // TODO: 添加 updater 插件时需要在 tauri.conf.json 中配置 plugins.updater
     // .plugin(tauri_plugin_updater::Builder::new().build())
     .invoke_handler(tauri::generate_handler![
@@ -38,7 +39,8 @@ pub fn run() {
         sftp::sftp_resume,
         sftp::sftp_cancel,
         sftp::sftp_move,
-        sftp::sftp_stat
+        sftp::sftp_stat,
+        sftp::sftp_read_text_file
     ])
     .setup(|app| {
       if cfg!(debug_assertions) {